Eliminate the one Unix emit, and one of three Win32 emits.
  The other two Win32 emits, setting the des from filedes/socketdes
  [HANDLE -> int] shouldn't be cast around, and we will likely move
  this code from poll/unix anyways.  The other emit, FD_SET errors,
  are gone in the latest win32 clib implementations, and again, our
  own win32 platform code will evaporate those problems.


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@63681 13f79535-47bb-0310-9956-ffa450edef68

+++ poll/unix/poll.c
@@ -162,8 +162,8 @@ APR_DECLARE(apr_status_t) apr_poll(apr_pollfd_t *aprset, int num, apr_int32_t *n
         tvptr = NULL;
     }
     else {
-        tv.tv_sec = apr_time_sec(timeout);
-        tv.tv_usec = apr_time_usec(timeout);
+        tv.tv_sec = (long)apr_time_sec(timeout);
+        tv.tv_usec = (long)apr_time_usec(timeout);
         tvptr = &tv;
     }
